Rename sexpr to sfsexp
The name sexpr exists mostly for hysterical reasons. Rename sexpr to
sfsexp where it names the project: in the README files and in
`configure.ac` In particular, `make dist-gzip` and the like will
generate realease tarballs `sfsexp-$VERSION.$SUFFIX` (with top level
directory `sfsexp-$VERSION`), in accordance with the project name and
the GitHub generated tarballs.

sexpr is left in old code as well where it denotes an actual
s-expression.