few fixes to run test_builds.sh
Ideally, the shebangs shouldn't be hardcoded since users might not have
the program available there, hence rely on /usr/bin/env (that will also
pick the preferred version, useful in the case of Python virtual envs
for example).

Fix the test invocation so libtool executes valgrind on the compiled
test and not the wrapper script to avoid a bunch of unrelated errors.